fixed bugs to do with stacked trace and fixed file input range for scan
git-svn-id: file:///afs/psi.ch/project/sls_det_software/svn/slsDetectorGui@22 af1100a4-978c-4157-bff7-07162d2ba061

+++ slsDetectorGui/include/qTabActions.h
@@ -77,7 +77,7 @@ private:
 	QCheckBox 	*chkSeparate;
 	QCheckBox 	*chkReturn;
 
-	float		*positions;
+	double		*positions;
 	QPalette	normal;
